% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{appshinyts}
\alias{appshinyts}
\title{Ejecutar aplicación Shiny de análisis de series de tiempo}
\usage{
appshinyts()
}
\value{
none
}
\description{
Aplicación que permite analizar series de tiempo  y realizar modelamiento con series
que contienen tendencia Lineal, Cuadrática o Cúbica.
Estacionalidad, patrones modelables como AR, MA, ARMA, ARIMA y Holt-Winters.
}
