% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{GroupTs}
\alias{GroupTs}
\title{Consolidar series de tiempo ajustadas AR, MA, ARMA, ARIMA, Holt-Winters}
\usage{
GroupTs(TsTrans, TsTend, TsSea, TsMod, TendRad, SeaRad, ModRad)
}
\arguments{ TsTrans: Transformación realizada a la serie de tiempo.
TsTrend: Recibe el ajuste realizado a la tendencia de la serie de tiempo.
TsSea: Modelamiento de la estacionalidad.
TenRad: Esta función trae la opción marcada por el usuario para la tendencia.
SeaRad: Esta función trae la opción marcada por el usuario para la estacionalidad.
ModRad: Esta función trae la opción marcada por el usuario para el modelo.
Ciclos: Recibe el modelo identificado para los ciclos.

\item{mtyp, }{xts, Time}
}
\value{
A matrix of the infile
}
\description{
Esta función recibe el ajuste realizado a la tendencia, a la estacionalidad y a los errores, los agrupa y muestra el ajuste total realizado.
}
