% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{grtsmod}
\alias{grtsmod}
\title{Grafica el desempeño del modelado de la serie de tiempo como AR, MA, ARMA, ARIMA, Holt-Winters}
\usage{
grtsmod(mod, xts, Time)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Grafica el desempeño del modelado de la serie
}
