% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{TsModel}
\alias{TsModel}
\title{TsModel}
\usage{
TsModel(mtyp, xts)
}
\arguments{
\item{mtyp}{es una entrada numérica para la componente estacional}
\item{xts}{es la serie de los datos previa al modelamiento}
}
\value{
A matrix of the infile
}
\description{
Modelar la serie de tiempo como AR, MA, ARMA, ARIMA, Holt-Winters. devuelve una serie de tiempo
residual
}
