% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{tsmodfordummy}
\alias{tsmodfordummy}
\title{tsmodfordummy}
\usage{
tsmodfordummy()
}
\arguments{
\item{Esta función no tiene argumentos}
}
\value{
A matrix of the infile
}
\description{
Esta función se usa como validación, para garantizar que la función se esté compilando adecuadamente. Genera un texto demostrativo
}
