% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{TsSeaModel}
\alias{TsSeaModel}
\title{TsSeaModel}
\usage{
TsSeaModel(mtyp, xts)
}
\arguments{
\item{mtyp}{es una entrada numérica para la componente estacional}
\item{xts}{es la serie de los datos previa al modelamiento}
}
\value{
A matrix of the infile
}
\description{
Modela una tendencia estacional, devuelve una serie de tiempo.
residual que puede ser modelada en AR, MA, ARMA, ARIMA, Holt-Winters
}
