% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{TsTrans}
\alias{TsTrans}
\title{Modelar la tendencia de la serie de tiempo}
\usage{
TsTrans(mtyp, xts)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Modela una tendencia Lineal, Cuadrático o Cúbica, devuelve una serie de tiempo.
residual que puede ser modelada en estacionalidad
}
