% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{appshinyts}
\alias{appshinyts}
\title{Ejecutar aplicación Shiny de análisis de series de tiempo}
\usage{
appshinyts()
}
\value{
none
}
\description{
Esta función permite ejecutar la aplicación de Shiny para el análisis estadístico de series de tiempo. Esta permite realizar modelamiento de series con tendencia lineal, cuadrática o cúbica, permite modelar la estacionalidad e identificar patrones como AR, MA, ARMA, ARIMA y Holt.Winters.
}
