% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{GroupTs}
\alias{GroupTs}
\title{Consolidar series de tiempo ajustadas AR, MA, ARMA, ARIMA, Holt-Winters}
\usage{
GroupTs(TsTrans, TsTend, TsSea, TsMod, TendRad, SeaRad, ModRad)
}
\arguments{
\item{mtyp, }{xts, Time}
}
\value{
A matrix of the infile
}
\description{
Modela la serie de tiempo, devuelve una serie de tiempo
residual.
}
