% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{TsModel}
\alias{TsModel}
\title{Modelar la serie de tiempo como AR, MA, ARMA, ARIMA, Holt-Winters}
\usage{
TsModel(mtyp, xts)
}
\arguments{
\item{mtyp, }{xts, Time}
}
\value{
A matrix of the infile
}
\description{
Modela la serie de tiempo, devuelve una serie de tiempo
residual.
}
