% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{tsmodfordummy}
\alias{tsmodfordummy}
\title{tsmodfordummy}
\usage{
tsmodfordummy()
}
\arguments{

}
\value{
A matrix of the infile
}
\description{
Esta funcion se usa como validacion, para garantizar que la funcion se esta compilando adecuadamente. Genera un texto demostrativo
}
