% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmodfor_01.R
\name{TsSeaModel}
\alias{TsSeaModel}
\title{Modelar la estacionalidad de la serie de tiempo}
\usage{
TsSeaModel(mtyp, xts)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Modela una tendencia estacional, devuelve una serie de tiempo.
residual que puede ser modelada en AR, MA, ARMA, ARIMA, Holt-Winters
}
